% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\docType{data}
\name{data_GSE17054}
\alias{data_GSE17054}
\title{This is data to be included in my package}
\format{a data frame with 4738 rows and 9 columns. 
The rows represent the genes and the columns represent the samples.}
\usage{
data(GSE17054)
}
\description{
Please see the package vignette for description.
}
\references{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE17054}
}
\keyword{dataset}

