% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\docType{data}
\name{data_GSE33223}
\alias{data_GSE33223}
\title{This is data to be included in my package}
\format{a data frame with 4114 rows and 30 columns. 
The rows represent the genes and the columns represent the samples.}
\usage{
data(GSE33223)
}
\description{
Please see the package vignette for description.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE33223}
}
\keyword{dataset}

