% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\docType{data}
\name{group_GSE57194}
\alias{group_GSE57194}
\title{This is data to be included in my package}
\format{a vector that represents the sample grouping for data_GSE57194. 
The elements of \emph{group_GSE57194} are either 'c' (control) or 'd' (disease).}
\usage{
data(GSE57194)
}
\description{
Please see the package vignette for description.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE57194}
}
\keyword{dataset}

