% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\docType{data}
\name{GSE17054}
\alias{GSE17054}
\alias{data_GSE17054}
\alias{group_GSE17054}
\title{Gene expression dataset GSE17054 from Majeti et al.}
\format{data_GSE17054 is a data frame with 4738 rows and 9 columns. 
The rows represent the genes and the columns represent the samples.

group_GSE17054 is a vector that represents the sample grouping for data_GSE17054. 
The elements of \emph{group_GSE17054} are either 'c' (control) or 'd' (disease).}
\source{
Obtained from \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE17054}
}
\usage{
data(GSE17054)
}
\description{
This dataset consists of 5 acute myeloid leukemia and 
4 control samples. The data frame data_GSE17054 includes the expression data 
while the vector group_GSE17054 includes the grouping information.
}
\references{
Majeti et al. Dysregulated gene expression networks in human acute 
myelogenous leukemia stem cells. Proceedings of the National Academy of 
Sciences, 106(9):3396-3401, 2009.
}
\keyword{dataset}

