% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\docType{data}
\name{GSE33223}
\alias{GSE33223}
\alias{data_GSE33223}
\alias{group_GSE33223}
\title{Gene expression dataset GSE33223 from Bacher et al.}
\format{data_GSE33223 is a data frame with 4114 rows and 30 columns. 
The rows represent the genes and the columns represent the samples.

group_GSE33223 is a vector that represents the sample grouping for 
data_GSE33223. The elements of \emph{group_GSE33223} are either 
'c' (control) or 'd' (disease).}
\source{
Obtained from \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE33223}
}
\usage{
data(GSE33223)
}
\description{
This dataset consists of 20 acute myeloid leukemia and 
10 control samples. The data frame data_GSE33223 includes the expression data 
while the vector group_GSE33223 includes the grouping information.
}
\references{
Bacher et al. Multilineage dysplasia does not influence prognosis in 
CEBPA-mutated AML, supporting the WHO proposal to classify these patients 
as a unique entity. Blood, 119(20):4719-22, 2012.
}
\keyword{dataset}

