% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\docType{data}
\name{GSE42140}
\alias{GSE42140}
\alias{data_GSE42140}
\alias{group_GSE42140}
\title{The gene expression dataset GSE42140 obtained from Gene Expression Omnibus}
\format{data_GSE42140 is a data frame with 4114 rows and 31 columns. 
The rows represent the genes and the columns represent the samples.

group_GSE42140 is a vector that represents the sample grouping for data_GSE42140. 
The elements of \emph{group_GSE42140} are either 'c' (control) or 'd' (disease).}
\usage{
data(GSE42140)
}
\description{
This dataset consists of 26 acute myeloid leukemia and 
5 control samples. The data frame data_GSE42140 includes the expression data 
while the vector group_GSE42140 includes the grouping information.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE42140}
}
\keyword{dataset}

