% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\docType{data}
\name{GSE57194}
\alias{GSE57194}
\alias{data_GSE57194}
\alias{group_GSE57194}
\title{Gene expression dataset GSE57194 from Abdul-Nabi et al.}
\format{data_GSE57194 is a data frame with 4114 rows and 12 columns. 
The rows represent the genes and the columns represent the samples.

group_GSE57194 is a vector that represents the sample grouping for data_GSE57194. 
The elements of \emph{group_GSE57194} are either 'c' (control) or 'd' (disease).}
\source{
Obtained from \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE57194}
}
\usage{
data(GSE57194)
}
\description{
This dataset consists of 6 acute myeloid leukemia and 
6 control samples. The data frame data_GSE57194 includes the expression data 
while the vector group_GSE57194 includes the grouping information.
}
\references{
Abdul-Nabi et al. In vitro transformation of primary human CD34+ cells by 
AML fusion oncogenes: early gene expression profiling reveals possible drug 
target in AML. PLoS One, 5(8):e12464, 2010.
}
\keyword{dataset}

