% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{pORACalc}
\alias{pORACalc}
\title{Over-representation Analysis}
\usage{
pORACalc(geneSet, DEGenes, measuredGenes, minSize = 0)
}
\arguments{
\item{geneSet}{a vector of gene names belong to the geneset}

\item{DEGenes}{a vector of differential expressed genes}

\item{measuredGenes}{a vector of all genes in the analysis}

\item{minSize}{the minimum number of DE genes in the geneSet}
}
\value{
p-value
}
\description{
Calculate p-value for over-representation Analysis
}
