% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampler}
\alias{sampler}
\title{sampler}
\usage{
sampler(iterations, vy_in, mX_in, prior, modelprior, modelpriorvec_in = NULL,
  cores = 1L)
}
\value{
The object returned is a list containing:
}
\description{
sampler
}
