% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysis.R
\name{splitS}
\alias{splitS}
\title{Spliting a vector of strings}
\usage{
splitS(x, splitSize = 5)
}
\arguments{
\item{x}{a vector of string}

\item{splitSize}{size of split vectors}
}
\value{
a list of string vectors
}
\description{
Split a vector of strings into several vectors
}
\details{
The function splits a vector of strings into several vectors
}
\examples{
x=paste("sample", seq(20))
splitS((x))
}
\author{
Tin Nguyen and Sorin Draghici
}

